gates/paypal/pyproject.toml
[project]
name = "paypal-processor-bot"
version = "0.1.0"
description = "PayPal card checker Telegram bot"
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = ["requests>=2.31.0", "python-telegram-bot>=20.0"]
